+++ GyroData/Source/Model/MotionCoordinate.swift
@@ -1,0 +1,12 @@
+//
+//  MotionCoordinate.swift
+//  GyroData
+//
+//  Created by 이태영 on 2023/02/03.
+//
+
+struct MotionCoordinate {
+    let x: Double
+    let y: Double
+    let z: Double
+}
